Move brokerage to new folder structure
